s5p: keep compatibility with utaformatix-generated ones

+++ libresvip/plugins/s5p/model.py
@@ -17,7 +17,7 @@ from libresvip.model.point import PointList
 
 
 class S5pPoint(NamedTuple):
-    offset: int
+    offset: float
     value: float
